requirements.txt
customtkinter
yt-dlp
spotdl
spotipy
requests
pillow